move temperature reading and storage to uint16 and store it mulitplied by 100 so 32.52 is 3252, maintain 8 bit reporting in user facing apis and reporting for now

+++ src/Pinoccio.h
@@ -38,7 +38,7 @@ class PinoccioClass {
 
     void goToSleep(uint32_t sleepForMs);
 
-    int8_t getTemperature();
+    int16_t getTemperature();
     int8_t getTemperatureOffset(void);
     void setTemperatureOffset(int8_t offset);
     void enableExternalAref();
